refactor: update src imports using index.ts
